backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
channels
channels_redis
redis-server
daphne